[USBVIDEO] - Provide KSCATEGORY_VIDEO and KSCATEGORY_CAPTURE as category - Queue 2 urbs for MJPEG capture and use the same frame context - Probe min/max and apply the values for probecommit - frame rate increases, image artifacts present

+++ drivers/usb/usbvideo/usbvideo.h
@@ -29,7 +29,7 @@ typedef struct _VS_PROBE_COMMIT_CONTROL
     ULONG   dwMaxVideoFrameSize;
     ULONG   dwMaxPayloadTransferSize;
     // UVC 1.1+
-    DWORD  dwClockFrequency;
+    ULONG   dwClockFrequency;
     UCHAR   bmFramingInfo;
     UCHAR   bPreferedVersion;
     UCHAR   bMinVersion;
